refactor: simplify `ape console` a bit
